Create travel time surface(s)

Create travel time surface(s)

Usage

create_travel_surface(
  friction_surface,
  points,
  extent_file = NA,
  id_col = NA,
  x_col = "x",
  y_col = "y",
  clip_flag = FALSE,
  transition_matrix_exists_flag = FALSE,
  save_transition_matrix = FALSE,
  individual_surfaces = FALSE,
  output_dir = NA,
  parallel = TRUE,
  cores = floor(parallel::detectCores()/2),
  check_existing = FALSE,
  overwrite = FALSE
)

Arguments

friction_surface

A raster containing the travel cost per pixel.

points

A dataframe containing fields for LABEL, X, and Y for point-level features.

extent_file

A raster or shapefile used to define extent of points and friction surface.

id_col

A character string corresponding to the column in points dataframe for LABEL field.

x_col

A character string corresponding to the column in points dataframe for x coordinate field.

y_col

A character string corresponding to the column in points dataframe for y coordinate field.

clip_flag

TRUE/FALSE: Should the friction surface extent be clipped?

transition_matrix_exists_flag

TRUE/FALSE: Has the already been constructed, must be in output_dir.

save_transition_matrix

TRUE/FALSE: Save transition matrix object? Useful for large areas.

individual_surfaces

TRUE/FALSE: Should individual travel surface be created for each point?

output_dir

A character string for output directory (rasters and intermediate outputs).

parallel

TRUE/FALSE: Should individual surfaces be created using parallel processing?

cores

an integer defining the number of CPU cores used for parallel processing

check_existing

TRUE/FALSE: Should rasters with LABEL in the output_dir be ignored?

overwrite

TRUE/FALSE: Should rasters in output_dir be overwritten?